move /pages into src
