+++ src/App.jsx
@@ -8,21 +8,19 @@ Change History:
           Remove unused imports (Lint)                
 *************************************************** */
 
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Landing from '../pages/Landing.jsx';
-import Carousel from '../pages/Carousel.jsx';
-import ImageGrid from '../pages/ImageGrid.jsx';
-
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import Carousel from "./pages/Carousel.jsx";
+import ImageGrid from "./pages/ImageGrid.jsx";
+import Landing from "./pages/Landing.jsx";
 
 export default function App() {
   return (
     <BrowserRouter>
-          <Routes>
-            <Route path='/' element={<Landing />} />
-            <Route path='/Carousel' element={<Carousel />} />
-            <Route path='/ImageGrid' element={<ImageGrid />} />
-          </Routes>
+      <Routes>
+        <Route path="/" element={<Landing />} />
+        <Route path="/Carousel" element={<Carousel />} />
+        <Route path="/ImageGrid" element={<ImageGrid />} />
+      </Routes>
     </BrowserRouter>
-
   );
 }
